refactor(datepicker): clean up date selection model
- Removes some methods from the date selection model that were put in place under the assumption that they'd be used, but we never ended up using them.
- Moves the selection model out of `core` and into `datepicker`. I decided to do it, because it seems very tightly coupled to the datepicker and I don't see us needing it in other components in the future.
