+++ src/material/core/datetime/index.ts
@@ -17,7 +17,6 @@ export * from './date-adapter';
 export * from './date-formats';
 export * from './native-date-adapter';
 export * from './native-date-formats';
-export * from './date-selection-model';
 
 
 @NgModule({
