+++ src/material/datepicker/public-api.ts
@@ -29,5 +29,6 @@ export * from './month-view';
 export * from './year-view';
 export * from './date-range-input';
 export * from './date-range-picker';
+export * from './date-selection-model';
 export {MatStartDate, MatEndDate} from './date-range-input-parts';
 export {MatMultiYearView, yearsPerPage, yearsPerRow} from './multi-year-view';
